Fixed: All compilation errors in dagger.
* Properly included the migration module based on the build variant via dagger.
