+++ app/build.gradle.kts
@@ -130,6 +130,9 @@ dependencies {
   androidTestImplementation(Libs.leakcanary_android_instrumentation)
   testImplementation(Libs.kotlinx_coroutines_test)
   testImplementation(Libs.TURBINE_FLOW_TEST)
+  debugImplementation(project(":defaultmigration"))
+  releaseImplementation(project(":defaultmigration"))
+  add("playStoreImplementation", project(":objectboxmigration"))
 }
 tasks.register("generateVersionCodeAndName") {
   val file = File("VERSION_INFO")
