+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/KiwixComponent.kt
@@ -19,23 +19,33 @@
 package org.kiwix.kiwixmobile.di.components
 
 import dagger.Component
+import org.kiwix.kiwixmobile.core.data.ObjectBoxDataMigrationHandler
 import org.kiwix.kiwixmobile.core.di.components.CoreComponent
+import org.kiwix.kiwixmobile.migration.di.module.MigrationModule
 import org.kiwix.kiwixmobile.di.KiwixScope
 import org.kiwix.kiwixmobile.di.components.ServiceComponent.Builder
 import org.kiwix.kiwixmobile.di.modules.JNIModule
 import org.kiwix.kiwixmobile.di.modules.KiwixModule
 import org.kiwix.kiwixmobile.di.modules.KiwixViewModelModule
+import org.kiwix.kiwixmobile.migration.di.module.DatabaseModule
 import org.kiwix.kiwixmobile.storage.StorageSelectDialog
 import org.kiwix.kiwixmobile.zimManager.OnlineLibraryManager
 
 @KiwixScope
 @Component(
   dependencies = [CoreComponent::class],
-  modules = [KiwixViewModelModule::class, KiwixModule::class, JNIModule::class]
+  modules = [
+    KiwixViewModelModule::class,
+    KiwixModule::class,
+    JNIModule::class,
+    MigrationModule::class,
+    DatabaseModule::class
+  ]
 )
 interface KiwixComponent {
   fun activityComponentBuilder(): KiwixActivityComponent.Builder
   fun serviceComponent(): Builder
   fun inject(storageSelectDialog: StorageSelectDialog)
   fun providesOnlineLibraryManager(): OnlineLibraryManager
+  fun provideObjectBoxDataMigrationHandler(): ObjectBoxDataMigrationHandler
 }
