+++ core/src/main/java/org/kiwix/kiwixmobile/core/main/CoreMainActivity.kt
@@ -47,7 +47,6 @@ import org.kiwix.kiwixmobile.core.CoreApp
 import org.kiwix.kiwixmobile.core.R
 import org.kiwix.kiwixmobile.core.base.BaseActivity
 import org.kiwix.kiwixmobile.core.base.FragmentActivityExtensions
-import org.kiwix.kiwixmobile.core.data.ObjectBoxDataMigrationHandler
 import org.kiwix.kiwixmobile.core.di.components.CoreActivityComponent
 import org.kiwix.kiwixmobile.core.downloader.DownloadMonitor
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.APP_NAME_KEY
@@ -165,9 +164,6 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
   abstract val appName: String
 
   @Inject
-  lateinit var objectBoxDataMigrationHandler: ObjectBoxDataMigrationHandler
-
-  @Inject
   lateinit var downloadMonitor: DownloadMonitor
 
   @Suppress("InjectDispatcher")
@@ -190,10 +186,6 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
     }
 
     setMainActivityToCoreApp()
-    // run the migration on background thread to avoid any UI related issues.
-    CoroutineScope(Dispatchers.IO).launch {
-      objectBoxDataMigrationHandler.migrate()
-    }
     lifecycleScope.launch(Dispatchers.IO) {
       createApplicationShortcuts()
     }
