+++ defaultmigration/src/main/AndroidManifest.xml
