+++ objectboxmigration/objectbox-models/default.json.bak
@@ -99,35 +99,6 @@
       "relations": []
     },
     {
-      "id": "4:6278838675135543734",
-      "lastPropertyId": "4:8812214350305159407",
-      "name": "LanguageEntity",
-      "properties": [
-        {
-          "id": "1:7795244654012809404",
-          "name": "id",
-          "type": 6,
-          "flags": 1
-        },
-        {
-          "id": "2:9116495537035444904",
-          "name": "locale",
-          "type": 9
-        },
-        {
-          "id": "3:452531964346972307",
-          "name": "active",
-          "type": 1
-        },
-        {
-          "id": "4:8812214350305159407",
-          "name": "occurencesOfLanguage",
-          "type": 5
-        }
-      ],
-      "relations": []
-    },
-    {
       "id": "5:3222423958972105425",
       "lastPropertyId": "11:6493253695033019145",
       "name": "HistoryEntity",
@@ -324,7 +295,8 @@
     349148274283701276,
     7257718270326155947,
     7394649290555378565,
-    8093454424037540087
+    8093454424037540087,
+    6278838675135543734
   ],
   "retiredIndexUids": [
     1293695782925933448,
@@ -399,7 +371,11 @@
     5555873126720275555,
     2724607601244650879,
     5485468735259326535,
-    4272820830206771469
+    4272820830206771469,
+    7795244654012809404,
+    9116495537035444904,
+    452531964346972307,
+    8812214350305159407
   ],
   "retiredRelationUids": [],
   "version": 1
