--- objectboxmigration/src/main/java/org/kiwix/kiwixmobile/objectboxmigration/di/module/ObjectBoxMigrationModule.kt
@@ -1,46 +1,0 @@
-/*
- * Kiwix Android
- * Copyright (c) 2025 Kiwix <android.kiwix.org>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package org.kiwix.kiwixmobile.objectboxmigration.di.module
-
-import dagger.Module
-import dagger.Provides
-import org.kiwix.kiwixmobile.core.data.ObjectBoxDataMigrationHandler
-import org.kiwix.kiwixmobile.objectboxmigration.ObjectBoxMigrationHandler
-import org.kiwix.kiwixmobile.objectboxmigration.data.ObjectBoxToLibkiwixMigrator
-import org.kiwix.kiwixmobile.objectboxmigration.data.ObjectBoxToRoomMigrator
-import org.kiwix.kiwixmobile.objectboxmigration.di.MigrationScope
-
-@Module
-class ObjectBoxMigrationModule {
-  @Provides
-  @MigrationScope
-  fun provideObjectBoxToLibkiwixMigrator() = ObjectBoxToLibkiwixMigrator()
-
-  @Provides
-  @MigrationScope
-  fun provideObjectBoxToRoomMigrator() = ObjectBoxToRoomMigrator()
-
-  @Provides
-  @MigrationScope
-  fun provideMigrationHandler(
-    objectBoxToRoomMigrator: ObjectBoxToRoomMigrator,
-    objectBoxToLibkiwixMigrator: ObjectBoxToLibkiwixMigrator
-  ): ObjectBoxDataMigrationHandler =
-    ObjectBoxMigrationHandler(objectBoxToRoomMigrator, objectBoxToLibkiwixMigrator)
-}
